More work on the new GridPage GUI.
Added Elektrothic font.
Improved TomThumb for readibility and aethetics.

Changed Tempo from int to float.

+++ avr/cores/megacommand/Midi/MidiClock.h
@@ -71,7 +71,7 @@ public:
 	
 	volatile bool updateSmaller;
 	uint16_t pll_x;
-	uint16_t tempo;
+	float tempo;
 	//bool transmit;
 	bool transmit_uart1;
     bool transmit_uart2;
